fix: ModelPickerGrid 메모리 폭증 — 썸네일 컴포넌트 공통화 (#258)
ModelPickerGrid 가 50+ 카드의 Civitai 썸네일을 lazy 없이 한꺼번에 로딩해
브라우저 메모리 폭증 / 크래시. LoRA 쪽에서 이미 해결한 패턴 (lazy img +
hover-play video) 을 공용 컴포넌트로 추출해 양쪽 적용.

## 신규 공용 컴포넌트
`frontend/src/components/common/MetadataMediaThumbnail.js`
- 이미지: `loading="lazy"` — viewport 밖 카드는 fetch 안 함
- 비디오: `preload="metadata"` 만 받고 hover 시에만 재생 (idle 메모리 최소화)
- `isVideoMedia(image)` helper export (LoraListModal 의 기존 helper 와 동일 로직)
- 본체 자체도 `React.memo` (image url 변경 안 되면 re-render 없음)

## 사용처 교체
- `LoraListModal`: 인라인 `LoraThumbnail` + `isVideoMedia` 제거,
  `MetadataMediaThumbnail` 으로 import 및 alias 유지 (기존 호출부 호환).
  LoraCard 자체에 `React.memo` 추가
- `ModelPickerGrid`: 인라인 `ModelThumbnail` 제거 (lazy 누락이었던 원인).
  `MetadataMediaThumbnail` 사용. 비디오 썸네일도 자동으로 hover-play 처리.
  `ModelCard` 에 `React.memo` 추가

## 영향
- LoRA: 회귀 없음 (logic 동일, 추출만)
- Model: 메모리 사용량 큰 폭 감소 — 50 카드 기준 즉시 로드되던 5MB+
  이미지/비디오가 viewport 진입 시에만 fetch
- 두 컴포넌트가 공유 파일을 사용하므로 향후 NSFW 처리 / 비디오 재생 정책
  변경 시 한 곳만 수정

## 후속 (선택)
- 더 큰 unification (전체 카드 그리드 / 검색 / 페이지네이션 공통화) 은
  별도 PR — LoraListModal 의 prompt 삽입 모드 / 다중 add 같은 LoRA 특화
  로직과 ModelPickerGrid 의 단일 선택 흐름이 다르므로 신중히 분리 필요

Closes #258
